@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,7 +22,9 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.5
+  shared_preferences: ^2.2.0
+  google_fonts: ^5.1.0
 
 dev_dependencies:
   flutter_test:
@@ -42,6 +44,9 @@ flutter:
   assets:
     - assets/images/mastercard.png
     - assets/images/contact_less.png
+    - assets/images/onBoarding1.png
+    - assets/images/onBoarding2.png
+    - assets/images/onBoarding3.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
