@@ -1,18 +1,32 @@
-import './credit_cards_page.dart';
+import 'package:credit_card_project/home.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'credit_cards_page.dart';
+import 'onboard/onboard.dart';
 
-void main() => runApp(MyApp());
+int? isViewed;
+Future<void> main() async {
+  SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle.dark);
+  WidgetsFlutterBinding.ensureInitialized();
+  SharedPreferences prefs=await SharedPreferences.getInstance();
+  isViewed=prefs.getInt('onBoard');
+  runApp(const MyApp());
+}
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Credit Cards Project',
-      theme: ThemeData(fontFamily: 'Lato'),
       debugShowCheckedModeBanner: false,
-      home: Scaffold(
-        body: SafeArea(child: CreditCardsPage()),
+      theme: ThemeData(
+        scaffoldBackgroundColor: Colors.white
       ),
+      title: 'Testing App',
+      home: isViewed !=0 ? const OnBoard() :  Home(),
     );
   }
 }
+
